Add intial version of firewall feature
This patch provides an initial firewall configuration that provides
complete isolation of the cluster from the outside world. The only way
to access the cluster is through the load balancer on ports 6443 (api),
443 (https), 80 (http) and currently 22623 (ignition). Sadly, we can't
automatically apply these firewalls to hosts, since the hcloud terraform
module has a bootstrapping problem[1], so this has to be done afterwards
manually using the Webinterface or CLI. Since all instances are tagged
already, it should be trivial to apply the right firewall rules based on
the tag.

Note on 22623, exposing this port through the loadbalancer permentently
to the internet is not ideal and should be changed. However, I haven't
come up with a better solution. This will be sorted out in the long run.

[1]: https://github.com/hetznercloud/terraform-provider-hcloud/issues/336